+++ setup.php
@@ -57,6 +57,7 @@ function plugin_thold_install($upgrade = false) {
 	api_plugin_register_hook($plugin, 'api_device_save', 'thold_api_device_save', 'setup.php');
 	api_plugin_register_hook($plugin, 'host_edit_bottom', 'thold_host_edit_bottom', 'setup.php');
 	api_plugin_register_hook($plugin, 'device_threshold_autocreate', 'thold_device_autocreate', 'setup.php');
+	api_plugin_register_hook($plugin, 'device_template_change', 'thold_device_template_change', 'setup.php');
 
 	// Automation Hooks
 	api_plugin_register_hook($plugin, 'create_complete_graph_from_template', 'thold_create_graph_thold', 'setup.php');
@@ -1228,17 +1229,22 @@ function thold_page_head() {
 function thold_device_edit_pre_bottom() {
 	html_start_box(__('Associated Threshold Templates', 'thold'), '100%', false, '3', 'center', '');
 
-	$host_template_id = db_fetch_cell_prepared('SELECT host_template_id FROM host WHERE id = ?' ,array(get_request_var('id')));
+	$host_id = get_request_var('id');
 
-	$threshold_templates = db_fetch_assoc_prepared('SELECT ptdt.thold_template_id, tt.name
-		FROM plugin_thold_host_template AS ptdt
+	$host_template_id = db_fetch_cell_prepared('SELECT host_template_id
+		FROM host
+		WHERE id = ?',
+		array($host_id));
+
+	$threshold_templates = db_fetch_assoc_prepared('SELECT pth.thold_template_id, tt.name
+		FROM plugin_thold_host AS pth
 		INNER JOIN thold_template AS tt
-		ON tt.id=ptdt.thold_template_id
-		WHERE ptdt.host_template_id = ?
+		ON tt.id = pth.thold_template_id
+		WHERE pth.host_id = ?
 		ORDER BY name',
-		array($host_template_id));
+		array($host_id));
 
-	html_header(array(__('Name', 'thold'), __('Status', 'thold')));
+	html_header(array(__('Name', 'thold'), __('Status', 'thold')), 2);
 
 	$i = 1;
 	if (cacti_sizeof($threshold_templates)) {
@@ -1257,12 +1263,15 @@ function thold_device_edit_pre_bottom() {
 
 			form_alternate_row("tt$i", true);
 			?>
-				<td class='left'>
-					<strong><?php print $i;?>)</strong> <?php print html_escape($item['name']);?>
-				</td>
-				<td>
-					<?php print $exists;?>
-				</td>
+			<td class='left'>
+				<strong><?php print $i;?>)</strong> <?php print html_escape($item['name']);?>
+			</td>
+			<td>
+				<?php print $exists;?>
+			</td>
+			<td class='nowrap right'>
+				<span title='<?php print __esc('Delete Threshold Template Association');?>' class='deletequery fa fa-times' id='ttremove<?php print $item['thold_template_id'];?>' data-id='<?php print $item['thold_template_id'];?>'></span>
+			</td>
 			<?php
 			form_end_row();
 
@@ -1271,6 +1280,59 @@ function thold_device_edit_pre_bottom() {
 	} else {
 		print '<tr><td class="templateAdd" colspan="2"><em>' . __('No Associated Threshold Templates.', 'thold') . '</em></td></tr>';
 	}
+
+	$available_thold_templates = db_fetch_assoc_prepared('SELECT DISTINCT tt.id, tt.name
+		FROM thold_template AS tt
+		LEFT JOIN plugin_thold_host_template AS ptht
+		ON ptht.thold_template_id = tt.id
+		WHERE ptht.host_template_id = ?
+		AND tt.id NOT IN (SELECT thold_template_id FROM plugin_thold_host WHERE host_id = ?)',
+		array($host_template_id, $host_id));
+
+	?>
+	<tr class='odd'>
+		<td class='saveRow' colspan='3'>
+			<table>
+				<tr style='line-height:10px;'>
+					<td class='nowrap templateAdd' style='padding-right:15px;'>
+						<?php print __('Add Threshold Template');?>
+					</td>
+					<td class='noHide'>
+						<?php form_dropdown('thold_template_id', $available_thold_templates, 'name', 'id', '', '', '');?>
+					</td>
+					<td class='noHide'>
+						<input type='button' class='ui-button ui-corner-all ui-widget' value='<?php print __esc('Add');?>' id='add_tt' title='<?php print __esc('Add Threshold Template to Device');?>'>
+					</td>
+				</tr>
+			</table>
+			<script type='text/javascript'>
+			$(function() {
+				$('[id^="ttremove"]').click(function(data) {
+					scrollTop = $(window).scrollTop();
+					$.post(urlPath+'plugins/thold/thold_templates.php?action=tt_remove', {
+						thold_template_id: $(this).attr('data-id'),
+						host_id: $('#id').val(),
+						__csrf_magic: csrfMagicToken }).done(function(data) {
+						loadPageNoHeader('host.php?header=false&action=edit&id='+$('#id').val());
+						$(window).scrollTop(scrollTop);
+					});
+				});
+
+				$('#add_tt').click(function() {
+					scrollTop = $(window).scrollTop();
+					$.post(urlPath+'/plugins/thold/thold_templates.php?action=add_tt', {
+						host_id: $('#id').val(),
+						thold_template_id: $('#thold_template_id').val(),
+						__csrf_magic: csrfMagicToken }).done(function(data) {
+						loadPageNoHeader('host.php?header=false&action=edit&id='+$('#id').val());
+						$(window).scrollTop(scrollTop);
+					});
+				});
+			});
+			</script>
+		</td>
+	</tr>
+	<?php
 
 	html_end_box();
 }
@@ -1436,6 +1498,25 @@ function thold_device_template_top() {
 
 		exit;
 	}
+}
+
+function thold_device_template_change($data) {
+	$device_id = $data['device_id'];
+	$device_template_id = $data['device_template_id'];
+
+	$thold_templates = db_fetch_assoc_prepared('SELECT *
+		FROM plugin_thold_host_template
+		WHERE host_template_id = ?',
+		array($device_template_id));
+
+	if (cacti_sizeof($thold_templates)) {
+		foreach($thold_templates as $tt) {
+			db_execute_prepared('REPLACE INTO plugin_thold_host (host_id, thold_template_id) VALUES (?, ?)',
+			array($device_id, $tt['thold_template_id']));
+		}
+	}
+
+	return $data;
 }
 
 function thold_device_autocreate($host_id) {