+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * security: Fix potential security exposure with unserialize() function
 
+* feature#445: Allow removal of Threshold from a specific device ignoring device template
+
 * feature#497: Show acknowledgment status on Thresholds tab instead of log
 
 * feature#528: Allow Thold to Use the BCC in addition to the TO line for Notification Lists